🧐 Fix remaining sync issues, adjust schema (#58)
This change resolves remaining issues introduced by #57 and #53.

The schema has had some slight adjustments made necessary by changes to
MyPurdue as well as breaking changes to the database library detailed
[here](https://www.roji.org/postgresql-dotnet-timestamp-mapping):

- `Capacity`, `Enrolled`, `RemainingSpace`, `WaitListCapacity`,
`WaitListCount`, and `WaitListSpace` are **no longer available** on
`Section`. Recent changes to MyPurdue have made it not feasible to
synchronize these values. This is tracked in #56 for a potential future
solution.
- `StartDate`/`EndDate` on `Meeting`, `Section`, and `Term` is now
stored as
[`DateOnly`](https://learn.microsoft.com/en-us/dotnet/api/system.dateonly)
objects, and may be null
- `StartTime` on `Meeting` is stored as a
[`TimeOnly`](https://learn.microsoft.com/en-us/dotnet/api/system.timeonly)
object, and may be null
- All dates and times are local time with no offset or time zone
information.

Some offline conversion tables were introduced for pieces of information
CatalogSync no longer has access to, such as campus/building short
codes. Tracked in #54 and #55 for potential future solutions.