🐜 Fix syncing Courses with duplicate Number + Title (#79)
A bug was recently discovered where course data was missing for courses
that had identical titles and numbers, but different subjects: #76

As an example instance, in Spring 2025, MyPurdue lists `MA 41600` and
`STAT 41600`, both titled "Probability," however Purdue.io only shows
`MA 41600` with every CRN from both of the original courses grouped
under it.

The root cause is that CatalogSync caches courses indexed only by their
number and title, not by their subject. This causes any courses with an
identical number and title to be grouped together, even those from
different subjects.

This change modifies the index to key off of subject code in addition to
number and title, resolving the initial bug.

It also includes the logic needed to "move" classes that were
erroneously synced with the existing logic to the correct place.